+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sl-design-system/popover
 
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [[`3295fa8`](https://github.com/sl-design-system/components/commit/3295fa8a92a7b0284a422232884f5fef77aa8537), [`3295fa8`](https://github.com/sl-design-system/components/commit/3295fa8a92a7b0284a422232884f5fef77aa8537)]:
+  - @sl-design-system/shared@0.2.10
+
 ## 0.1.10
 
 ### Patch Changes